requirements.txt
langgraph==0.5.0
pydantic>=2.6,<3
PyYAML>=6.0
networkx>=3.2
requests>=2.31
# openai  # optional: install if you want real OpenAI calls
